Finished the Reheap function
Finished the Reheap function. The frequency_table can now be a Huffman table.

The next steps are:
Generating bitcodes
Encoding the table, and header data.
Writing the header to file, then writing the table to the file.

The header data includes:
The length of the file name.
The file name.
How many entries are in the Huffman table.