Writing modifications
All files needed to be open and written to as binary files. So I made that change.

Since we are opening as binary we can't use the out << output_byte; in Huff.

Changed it to use out.write((char*)&output_byte, sizeof(output_byte));

Changed puff to open and write as binary.